// src/types.ts
import { DataQuery, DataSourceJsonData } from '@grafana/data';

export enum AutoOffsetReset {
  EARLIEST = 'earliest',
  LATEST = 'latest',
}

export enum TimestampMode {
  Now = 'now',
  Message = 'message',
}

export type AutoOffsetResetInterface = {
  [key in AutoOffsetReset]: string;
};

/**export interface MyQuery extends DataQuery {
  queryText?: string;
  constant: number;
}**/

export const DEFAULT_QUERY: Partial<MyQuery> = {
  constant: 6.5,
};

/**
 * These are options configured for each DataSource instance
 */
export interface MyDataSourceOptions extends DataSourceJsonData {
  path?: string;
  useMTLS?: boolean
  bootstrapServers?: string;
  mtlsPrivateKeyPath?: string;
  mtlsPublicKeyPath?: string;
  mtlsKeyPassword?: string;
}

/**
 * Value that is used in the backend, but never sent over HTTP to the frontend
 */
export interface MySecureJsonData {
  mtlsKeyPassword?: string;
}

export interface MyQuery extends DataQuery {
  topicName: string;
  partition: number;
  withStreaming: boolean;
  autoOffsetReset: AutoOffsetReset;
  timestampMode: TimestampMode;
  constant: number;
}

export const defaultQuery: Partial<MyQuery> = {
  partition: 0,
  withStreaming: true,
  autoOffsetReset: AutoOffsetReset.LATEST,
  timestampMode: TimestampMode.Now,
  constant: 6.5,
};